Fix onCanvasClick when scrolled
When you scroll a page with the timeline, the onCanvasClick incorrectly determines the group/row for many browsers (Edge, Firefox, and Chrome 61). These browsers always return 0 for the body element's scrollTop property. Since the element.offsetParent bubbles up to the body element rather than the actual scrolling element, the onCanvasClick returns row/group as if the page was not scrolled. I mainly use chrome and didn't notice the issue till the latest chrome update.

Ideally offsetParent would return the scrolling element as the top level parent, but this was a quick solution that fixed the issue I was having. There may be better ways to solve. I tried to make it fall back to previous behavior if it didn't meet specific requirements.